refactor(webui): remove utils compatibility shims
