--- nanobot/utils/session_attachments.py
@@ -1,4 +1,0 @@
-"""Compatibility imports for WebUI session attachment helpers."""
-
-from nanobot.webui.attachments import *  # noqa: F401,F403
-
