--- nanobot/utils/webui_sidebar_state.py
@@ -1,3 +1,0 @@
-"""Compatibility imports for WebUI sidebar state helpers."""
-
-from nanobot.webui.sidebar_state import *  # noqa: F401,F403
