--- nanobot/utils/webui_thread_disk.py
@@ -1,4 +1,0 @@
-"""Compatibility imports for WebUI thread disk helpers."""
-
-from nanobot.webui.thread_disk import *  # noqa: F401,F403
-
