--- nanobot/utils/webui_transcript.py
@@ -1,4 +1,0 @@
-"""Compatibility imports for WebUI transcript helpers."""
-
-from nanobot.webui.transcript import *  # noqa: F401,F403
-
