--- nanobot/utils/webui_turn_helpers.py
@@ -1,4 +1,0 @@
-"""Compatibility imports for WebUI turn helpers."""
-
-from nanobot.webui.turn_helpers import *  # noqa: F401,F403
-
